lambda/app/requirements.txt
argon2-cffi ~= 20.1